+++ employees-service/src/main/java/com/example/employees_service/dto/EmployeeDto.java
@@ -21,7 +21,7 @@ public class EmployeeDto {
 
     private String patronymic;
 
-    private int stuffId;
+    private Long stuffId;
 
     private String employeePost;
